+++ services/web/src/features/suggestions/sections/SuggesedSections.tsx
@@ -1,5 +1,4 @@
-import { SuggestionStatus } from '@whitewater-guide/commons';
-import React, { useState } from 'react';
+import React from 'react';
 import { useQuery } from 'react-apollo';
 import { Loading } from '../../../components';
 import {
@@ -9,32 +8,26 @@ import {
 } from './suggestedSections.query';
 import SuggestedSectionsInfinite from './SuggestedSectionsInfinite';
 
-const DEFAULT_STATUSES = [SuggestionStatus.PENDING];
-
 export const SuggesedSections: React.FC = () => {
-  const [statusFilter, setStatusFilter] = useState(DEFAULT_STATUSES);
-
   const { data, loading, fetchMore } = useQuery<QResult, QVars>(
     SUGGESTED_SECTIONS_QUERY,
     {
       fetchPolicy: 'network-only',
       variables: {
-        filter: { status: statusFilter },
+        filter: { editable: true, verified: false },
       },
     },
   );
 
-  if (loading && !(data && data.suggestedSections)) {
+  if (loading && !(data && data.sections)) {
     return <Loading />;
   }
 
   return (
     <React.Fragment>
       <SuggestedSectionsInfinite
-        suggestedSections={data!.suggestedSections}
+        suggestedSections={data!.sections}
         fetchMore={fetchMore}
-        statusFilter={statusFilter}
-        setStatusFilter={setStatusFilter}
       />
     </React.Fragment>
   );